Adding tables(OB); fix time;

+++ SNTDataBase/src/Data/Payment/PaymentTabDS.cpp
@@ -11,7 +11,7 @@ namespace LM
 		IsDirtyDecorator(m_Data.Amount.DrawEdit(u8"Сумма платежа"));
 
 		ImGui::PushItemWidth(250);
-		IsDirtyDecorator(ImGui::Combo(u8"Форма платежа", &m_Data.FormOfPayment, PaymentData::s_FormOfPaymentString.data(), PaymentData::s_FormOfPaymentString.size()));
+		IsDirtyDecorator(ImGui::Combo(u8"Форма платежа", &m_Data.FormOfPayment, PaymentData::s_FormOfPaymentString.data(), (int)PaymentData::s_FormOfPaymentString.size()));
 		ImGui::PopItemWidth();
 
 		IsDirtyDecorator(InputTextString(u8"Номер платежа", m_Data.DocumentNumber, 250));
