+++ SNTDataBase/src/UI/Tabs/TabDataStruct/TabDataStruct.cpp
@@ -3,12 +3,12 @@
 namespace LM
 {
 
-	bool TabDataStruct::InputTextString(std::string_view _Name, std::string& _Data, const int _ItemWidth, const ImGuiInputTextFlags_ _Flag)
+	bool TabDataStruct::InputTextString(std::string_view _Name, std::string& _Data, const float _ItemWidth, const ImGuiInputTextFlags_ _Flag)
 	{
 		if (_ItemWidth)
 			ImGui::PushItemWidth(_ItemWidth);
 
-		int BufferSize = _Data.size() + 128;
+		size_t BufferSize = _Data.size() + 128;
 		char* Buffer = new char[BufferSize] { 0 };
 		memcpy(Buffer, _Data.data(), _Data.size());
